# Double query translation