--- a/src/NHibernate.Test/NHSpecificTest/GH1547/Entity.cs
+++ b/src/NHibernate.Test/NHSpecificTest/GH1547/Entity.cs
@@ -1,0 +1,10 @@
+﻿using System;
+
+namespace NHibernate.Test.NHSpecificTest.GH1547
+{
+	class Entity
+	{
+		public virtual Guid Id { get; set; }
+		public virtual string Name { get; set; }
+	}
+}