--- a/src/NHibernate.Test/NHSpecificTest/GH1547/Mappings.hbm.xml
+++ b/src/NHibernate.Test/NHSpecificTest/GH1547/Mappings.hbm.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<hibernate-mapping xmlns="urn:nhibernate-mapping-2.2" assembly="NHibernate.Test" namespace="NHibernate.Test.NHSpecificTest.GH1547">
+
+	<class name="Entity">
+		<id name="Id" generator="guid.comb" />
+		<property name="Name" />
+	</class>
+
+</hibernate-mapping>