… src/NHibernate.Test/Async/NHSpecificTest/GH1547/Fixture.cs