This is a kind of follow-up to #1544, for minimizing its potential impact on performances. But it also improves things for other cases.

(By the way, I have checked with the "raw data benchmark", #1544 was not having a measurable impact on it, even by tweaking it for using a Linq query with a non-parameterized constant. But the way it works, pulling a whole table without paging, the incurred IO are quite big and so it is not really surprising we cannot see the impact of a potential performance loss occurring only once per query.)

# First commit rationals:
_(Avoid re-translating NhLinqExpression)_

Whatever the querying API, executing a query currently involves two steps:
 * Create the query (prepare in LINQ)
 * Set parameters (not exposed to users with LINQ)
 * Execute the query

First and last steps need the query plan, and these steps ask it from the plan cache, which will generate it if it is missing.
The first step needs it notably to initialize its parameter list. The last step of course needs it for executing the query.

So if a plan is not cacheable (which since #1544 may happen with LINQ queries), it will forcibly be regenerated at the execution step, instead of being retrieved from cache.

At the first step, the generated plan cannot be stored in the query for later re-use, because the query or its execution conditions may be altered and may require another plan for the query execution. This occurs:
 * When a query use a list parameter. Such a parameter has to be replaced by as many parameters as it has values in its list. (Moreover in the LINQ case this query alteration needs the original query translation, and recomputes it.)
 * For non Linq queries, the user may enable session filters after having created a query. Query plan keys depend on the list of enabled filter.

But the query translation, done when generating a plan, can be re-used if the query expression instance is not changed for another one. So this PR alter `NhLinqExpression`, which was already storing its translation results, for yielding them again instead of recomputing them in case of later `Translate` call.

This change is not done for non-LINQ queries because they are `string` based, without a common object persisting from first plan look-up to second look-up. Anyway, they would less benefit of the change since non-LINQ query plans are currently always cacheable.

By the way this change lighten an old "re-translation" concern occurring with queries having list parameters, see code comments changes.

# Second commit:
_(Fix a wrong parameter value for query translation)_

The first commit, as you can read in its code comment, assumes a query is always translated for the same session factory and with the same `bool filter` parameter. This was not true for `filter` due to a latent bug causing this parameter to have a wrong value for some filter query case. But this bug was not having any impact because this parameter is currently unused.

The second commit just fix that, for having a cleaner code. It currently changes nothing to the feature behavior, so we could also just drop that commit.