@fredericDelaporte 2018-01-21
Typos fixed, thanks.

> Any tests required?

Besides running existing ones and seeing they still succeed, no additional ones.
With query plan there are logs allowing to checks how many generations or cache hit occurred. But there are none in the case of query translation, and I rather prefer not adding logs just for tests. There are also too many things involved in triggering the query translation for elaborating a NSubstitute based test.