@fredericDelaporte 2018-01-22
Cases which are re-using the translation were failing because the query plan compilation alters the supplied query translation (`AstNode` tree), although it also creates a new nodes tree. This happens in `HqlSqlTranslator.Translate()`. In it, it is obvious that the filter case always mutates the input. But moreover, it appears that `HqlSqlWalker.statement()` also mutate the input, although this is not obvious (and maybe not expected indeed). When looking at `AstNode.ToStringTree()`, we can see that the entity alias (`e` by example) gets replaced by the sql alias (`entity0_`) excepted in the from clause.
Fixing the filter case for generating a new tree instead of mutating the input would be easy, there are already similar cases for parameter expansion and polymorphic tree building.
But about the `HqlSqlWalker`, I get lost trying to find where it does mutate its input.

Since the `AstNode` returned by `NhLinqExpression.Translate` was previously never re-accessed from `NhLinqExpression` instances (neither inside them nor from using their properties), having it mutated was not having any impact. (Neither needed nor causing bug.) But it prevents re-using it inside this PR.

The more robust solution is then to yield a copy of the `AstNode` instead of the stored one. This may be a bit redundant in the polymorphic case, where the tree is always duplicated. So we could instead "hack" the polymorphic detector for yielding a duplicated tree in non-polymorphic cases too. But that is not its responsibility, so I have ruled out this option.

The other option would be to avoid mutating the input when building the query plan. But this looks like requiring many more code changes, with some cases which will anyway have to duplicate the whole `AstNode` tree (filter case).