@fredericDelaporte 2018-01-23
Although I think the "raw data access bench" is not very suitable for checking impact of this change, I have still run it: no significant differences in timings.