@hazzik 2018-01-23
>no significant differences in timings.

Still, better or worse?