@fredericDelaporte 2018-01-24
Below standard deviation, so cannot say. I will put the figures later. But It think a different bench is needed for really checking. Like one that runs queries having lightweight results in parallel and compute the average throughput.
On Linq queries not needing to recompute their plan, it may show a loss, while on queries recomputing their plan (not cacheable ones due to non-parameterized constant + those having a list parameter with a cardinality other than one) it may show a win. But as those may be smaller than the variability of the database response time, it is not even sure.
For better checking it should instead be a test using a driver mock yielding in-memory results. (In such case better no more run it parallel for minimizing impact of other OS's processes which may cause random competing load.)