@fredericDelaporte 2018-01-24
Test | NH prior to the PR | NH altered by this PR  | % (base is prior to the PR)
------------ | -------------|-------------|------------
Change tracking fetches, set fetches (50 runs), no caching | 1 369.71ms, Enum: 1.71ms, 206 155KB | 1 352.97ms, Enum: 1,61ms, 206 147KB | -1.3%ms, Enum: -5.8%, -0.0004%KB
Change tracking individual fetches (100 elements, 50 runs), no caching | 0.32ms, 64KB | 0.37ms, 64KB | +15.6%ms, 0%KB
Change tracking fetches, eager load fetches, 3-node split graph, 1000 root elements (50 runs), no caching | 179.38ms, 30 768KB | 178.94ms, 30 768KB | -2.5%ms, 0%KB

So that seems like this PR would be better for tests involving LINQ (first and third).
But, as the individual fetches (second line) does not involve any LINQ (it is implemented with an `ISession.Get`), it demonstrates the lack of reliability of comparing those figures when they are close. (Standard deviation is 2ms for this test, so even differences beyond standard deviation are not reliable.)