@fredericDelaporte 2018-01-25
I have added a test fixture (set as explicit) adequate for benchmarking the change.

Raw results:
```
With 1547:
Run 1
List parameter average time: 22,5247193ms (s 4,333056688069ms)
Non parameterized constant average time: 23,4214071ms (s 3,9677044980547ms)
Simple LINQ average time: 8,0586627ms (s 2,548815048447ms)

Run 2
List parameter average time: 22,426151ms (s 4,65178411170821ms)
Non parameterized constant average time: 23,1536621ms (s 3,53700987800864ms)
Simple LINQ average time: 7,8267019ms (s 2,90186192777228ms)

Without 1547:
Run 1
List parameter average time: 25,4970112ms (s 5,54136760824837ms)
Non parameterized constant average time: 35,9412004ms (s 4,5727023529384ms)
Simple LINQ average time: 7,7744208ms (s 2,72784471485745ms)

Run 2
List parameter average time: 25,2826661ms (s 5,19295881027821ms)
Non parameterized constant average time: 35,2281946ms (s 4,12219056603083ms)
Simple LINQ average time: 7,9850022ms (s 2,60721099540673ms)
```

So to compare things (run 2):
Test | NH prior to the PR | NH altered by this PR  | % (base is prior to the PR)
------------ | -------------|-------------|------------
List parameter | 25.28ms | 22.42ms | -11.3%ms
Non parameterized constant | 35.22ms | 23.15ms | -34.3%ms
Simple LINQ | 7.98ms | 7.82ms | -2%ms

So cases optimized by the change (the two first) are actually optimized, while the case for which it can only be a penalty is not significantly impacted. (The apparent gain it even has with the PR is below sample standard deviation, so it is not significant.)