@fredericDelaporte 2018-03-13
@hazzik, I think this PR should also be included in 5.1.