@hazzik 2018-03-14
Drop last commit. It's ugly. 

I still think that this can be done without need of additional clones. But I don't want to waste time _now_ exploring these options.